requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
greenlet==3.2.4

# Database
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10

# HTTP & Async
aiohttp==3.12.15
httpx==0.28.1

# Vector Database (Milvus)
pymilvus==2.6.0
sentence-transformers==5.1.0
torch==2.8.0
numpy>=1.17

# Graph Database (Neo4j)
neo4j==5.28.2

# LangChain & AI
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.8
langchain-experimental>=0.3.0
langgraph==0.6.6

# ML & NLP
transformers>=4.41.0
huggingface-hub>=0.20.0

# API & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# XML Processing
lxml==5.3.0

# Utilities
# PyYAML==6.0.2  # 사용하지 않음
# click==8.2.1  # 사용하지 않음

# Development & Testing
# pytest==8.3.5  # 테스트 코드 없음
# pytest-asyncio==0.24.0  # 테스트 코드 없음
# httpx-sse==0.4.0  # 사용하지 않음

# MIDI
midiutil==1.2.1
